added last of styles
