+++ packages/molly/src/lib/components/Molly.svelte
@@ -4,7 +4,7 @@
 	import { SSE } from 'sse.js';
 	import { onMount } from 'svelte';
 	import MollyButton from './MollyButton.svelte';
-	import MollyWindow from './MollyWindow.svelte';
+	import MollyHeader from './MollyHeader.svelte';
 	import MollyForm from './MollyForm.svelte';
 
 	let query: string = '';
@@ -80,7 +80,7 @@
 
 <MollyButton>
 	<div class="chat-container fixed bottom-0 right-0 w-80 bg-slate-200 dark:bg-slate-600">
-		<MollyWindow/>
+		<MollyHeader/>
 			<div class="h-full grid grid-rows-[1fr_auto]">
 					<div bind:this={chatWindow} class="messages-container h-80 bg-surface-500/30 overflow-y-auto">
 						{#each chatMessages as message}
